training/requirements.txt
ultralytics
opencv-python
scikit-learn
tensorflow-datasets>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm
